+++ src/model.h
@@ -4,11 +4,17 @@
 #include <Rcpp.h>
 #include <numeric>
 
+#include "dgCMatrix.h"
+
 using namespace Rcpp;
+
+// C++ interface to access elements of a POMDP model
 
 // NOTE: Episode in time-dependent POMDPs are currently unsupported.
 // NOTE: All indices are 0-based.
 
+
+// Access model information
 inline bool is_solved(const List& model) { 
   return model.containsElementNamed("solution");
 }
@@ -17,30 +23,7 @@ inline bool is_converged(const List& model) {
   return as<LogicalVector>(as<List>(model["solution"])["converged"])[0];
 }
 
-inline NumericMatrix transition_matrix(const List& model, int action) {
-  return as<NumericMatrix>(as<List>(model["transition_prob"])[action]);
-}
-
-inline double transition_prob(const List& model, int action, int start_state, int end_state) {
-  return transition_matrix(model, action)(start_state, end_state);
-}
-
-inline NumericMatrix observation_matrix(const List& model, int action) {
-  return as<NumericMatrix>(as<List>(model["observation_prob"])[action]);
-}
-
-inline double observation_prob(const List& model, int action, int end_state, int observation) {
-  return observation_matrix(model, action)(end_state, observation);
-}
-
-inline NumericMatrix reward_matrix(const List& model, int action, int start_state) {
-  return as<NumericMatrix>(as<List>(as<List>(model["reward"])[action])[start_state]);
-}
-
-inline double reward_val(const List& model, int action, int start_state, int end_state, int observation) {
-  return reward_matrix(model, action, start_state)(end_state, observation);
-}  
-
+// More accessors
 inline NumericVector start_vector(const List& model) {
   return as<NumericVector>(model["start"]);
 }  
@@ -92,6 +75,180 @@ inline DataFrame get_pg(const List& model, int epoch = 0) {
   epoch = get_pg_index_cpp(model, epoch);
   return as<DataFrame>(as<List>(as<List>(model["solution"])["pg"])[epoch]);
 }
+
+
+// Transitions & Observations
+// Can be a dense matrix or a dgCMatrix
+// Available functions are: x_matrix returns a dense matrix, x_prob returns double, and x_row returns a vector
+inline NumericMatrix transition_matrix(const List& model, int action, int episode = -1) {
+  RObject acts;
+  if (episode >= 0)
+    acts = as<List>(as<List>(model["transition_prob"])[episode])[action];
+  else
+    acts = as<List>(model["transition_prob"])[action];
+  
+  if (is<NumericMatrix>(acts)) 
+    return as<NumericMatrix>(acts); 
+  
+  // dgCMatrix
+  return dgCMatrix(as<S4>(acts)).dense();
+}
+
+inline double transition_prob(const List& model, int action, int start_state, int end_state, int episode = -1) {
+  RObject acts;
+  if (episode >= 0)
+    acts = as<List>(as<List>(model["transition_prob"])[episode])[action];
+  else
+    acts = as<List>(model["transition_prob"])[action];
+  
+  // dense matrix
+  if (is<NumericMatrix>(acts)) 
+    return as<NumericMatrix>(acts)(start_state, end_state); 
+  
+  // dgCMatrix
+  return dgCMatrix(as<S4>(acts)).at(start_state, end_state);
+}
+
+inline NumericVector transition_row(const List& model, int action, int start_state, int episode = -1) {
+  RObject acts;
+  if (episode >= 0)
+    acts = as<List>(as<List>(model["transition_prob"])[episode])[action];
+  else
+    acts = as<List>(model["transition_prob"])[action];
+  
+  // dense matrix
+  if (is<NumericMatrix>(acts)) 
+    return as<NumericMatrix>(acts).row(start_state); 
+  
+  // dgCMatrix
+  return dgCMatrix(as<S4>(acts)).row(start_state);
+}
+
+inline NumericMatrix observation_matrix(const List& model, int action, int episode = -1) {
+  RObject acts;
+  if (episode >= 0)
+    acts = as<List>(as<List>(model["observation_prob"])[episode])[action];
+  else
+    acts = as<List>(model["observation_prob"])[action];
+  
+  // dense matrix
+  if (is<NumericMatrix>(acts)) 
+    return as<NumericMatrix>(acts); 
+  
+  // dgCMatrix
+  return dgCMatrix(as<S4>(acts)).dense();
+}
+
+inline double observation_prob(const List& model, int action, int end_state, int observation, int episode = -1) {
+  RObject acts;
+  if (episode >= 0)
+    acts = as<List>(as<List>(model["observation_prob"])[episode])[action];
+  else
+    acts = as<List>(model["observation_prob"])[action];
+  
+  // dense matrix
+  if (is<NumericMatrix>(acts)) 
+    return as<NumericMatrix>(acts)(end_state, observation); 
+  
+  // dgCMatrix
+  return dgCMatrix(as<S4>(acts)).at(end_state, observation);
+}
+
+inline NumericVector observation_row(const List& model, int action, int end_state, int episode = -1) {
+  RObject acts;
+  if (episode >= 0)
+    acts = as<List>(as<List>(model["observation_prob"])[episode])[action];
+  else
+    acts = as<List>(model["observation_prob"])[action];
+  
+  // dense matrix
+  if (is<NumericMatrix>(acts)) 
+    return as<NumericMatrix>(acts).row(end_state); 
+  
+  // dgCMatrix
+  return dgCMatrix(as<S4>(acts)).row(end_state);
+}
+
+
+// Reward
+
+// TODO add support for episodes
+
+// Can be a dense matrix or a data.frame
+// Available are reward_matrix and reward_val
+inline NumericMatrix reward_matrix(const List& model, int action, int start_state, int episode = -1) {
+  RObject reward = model["reward"];
+  if (episode >= 0)
+    reward = as<List>(reward)[episode];
+  
+  if (is<DataFrame>(reward)) {
+    DataFrame df = as<DataFrame>(reward);
+    IntegerVector actions = df[0], start_states = df[1], end_states = df[2], observations = df[3];
+    NumericVector values = df[4]; 
+    
+    NumericMatrix rew(get_states(model).size(), get_obs(model).size());
+    
+    for (auto i = 0; i < df.nrows(); ++i) {
+      if(
+        (IntegerVector::is_na(actions[i]) || actions[i] == action) && 
+        (IntegerVector::is_na(start_states[i]) || start_states[i] == start_state)) {
+      
+          if (IntegerVector::is_na(end_states[i]) &&
+              IntegerVector::is_na(observations[i])) 
+                  std::fill(rew.begin(), rew.end(), values[i]);
+          else if (IntegerVector::is_na(end_states[i]))
+                  rew(_ , observations[i]) = NumericVector(rew.rows(), values[i]);
+          else if (IntegerVector::is_na(observations[i]))
+                  rew(end_states[i], _) = NumericVector(rew.cols(), values[i]);
+          else
+                  rew(end_states[i], observations[i]) = values[i];
+      }
+    }
+        
+    return rew;
+  }
+  
+  // it is a matrix
+  return as<NumericMatrix>(as<List>(as<List>(reward)[action])[start_state]);
+}
+
+
+// Note: R_index does not apply to episode!!!
+inline double reward_val(const List& model, int action, int start_state, int end_state, int observation,
+  int episode = -1, bool R_index = FALSE) {
+  RObject reward = model["reward"];
+  if (episode >= 0)
+    reward = as<List>(reward)[episode];
+  
+  // data.frame indices are 1-based!!!
+  if (!R_index) {
+    action++; start_state++; end_state++; observation++;
+  }
+  
+  if (is<DataFrame>(reward)) {
+    DataFrame df = as<DataFrame>(reward);
+    // find the best matching entry
+    IntegerVector actions = df[0], start_states = df[1], end_states = df[2], observations = df[3];
+    NumericVector values = df[4]; 
+    
+    for (auto i = df.nrows(); i >= 0; --i) {
+      if(
+          (IntegerVector::is_na(actions[i]) || actions[i] == action) && 
+          (IntegerVector::is_na(start_states[i]) || start_states[i] == start_state) &&
+          (IntegerVector::is_na(end_states[i]) || end_states[i] == end_state) &&
+          (IntegerVector::is_na(end_states[i]) || end_states[i] == observation)
+        )
+        return values[i];
+        
+    }
+    return 0.0;
+    
+  }
+    
+  // it is not a data.frame so it must be a List of matrices
+  return reward_matrix(model, action, start_state, episode)(end_state, observation);
+}  
+